Revert "Add runtime_ticks initialization and declare sys_getprocload syscall"
This reverts commit 80ce21b4314e141eecdc1be14b62f953d1e9f13a.

+++ kernel/proc.h
@@ -91,7 +91,7 @@ struct proc {
   int killed;                  // If non-zero, have been killed
   int xstate;                  // Exit status to be returned to parent's wait
   int pid;                     // Process ID
-  uint runtime_ticks;           // CPU time used by this process in timer ticks
+  uint64 runtime_ticks;        // CPU time used by this process in timer ticks
   int throttle_limit;          // Cooldown duration in ticks when throttled
   uint64 throttle_until;       // Global tick when this process may run again
